Modified SystemData.sol to not ingest planet orbit
Made it so planet orbits are evenly distributed instead of reading orbits from json input

+++ packages/hardhat/contracts/Structs.sol
@@ -10,11 +10,11 @@ library Structs {
 
   struct System {
     string name;
-    uint256 distToSol;
-    uint256 radius;
+    uint16 distToSol;
+    uint8 radius;
     string color;
     address owner;
     uint256[] planets;
   }
-  
+
 }
